requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1   # <-- add this
python-dotenv==1.0.1
numpy==1.26.4
chromadb==0.5.5
fastembed==0.6.1
pypdf==4.2.0
groq==0.11.0
openai==1.40.1
pypdfium2==4.30.0
pytesseract==0.3.10
Pillow==10.4.0
python-docx==1.1.0
openpyxl==3.1.5


